Refactored TestRunSummary and TestRunSummaryConcatenated into their own classes.
